@@ -26,7 +26,7 @@ namespace Userscripts.Controllers
         [HttpGet]
         public IEnumerable<WeatherForecast> Get()
         {
-            Random rng = new Random();
+            Random rng = new();
             return Enumerable.Range(1, 1).Select(index => new WeatherForecast
                 {
                     Date = DateTime.Now.AddDays(index),
