@@ -1,13 +1,8 @@
 ﻿using Userscripts.Models;
-using IdentityServer4.EntityFramework.Options;
 using Microsoft.AspNetCore.ApiAuthorization.IdentityServer;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Userscripts.Controllers;
+using Duende.IdentityServer.EntityFramework.Options;
 
 namespace Userscripts.Data
 {
